+++ packages/orchestrator-ui-components/src/components/WfoSubscription/utils/utils.spec.ts
@@ -1,12 +1,13 @@
 import { EuiThemeComputed } from '@elastic/eui';
 
-import { SubscriptionAction } from '../../../hooks';
 import {
     FieldValue,
     ProcessStatus,
+    SubscriptionAction,
     SubscriptionDetailProcess,
     WorkflowTarget,
-} from '../../../types';
+} from '@/types';
+
 import {
     flattenArrayProps,
     getFieldFromProductBlockInstanceValues,
