+++ packages/orchestrator-ui-components/src/hooks/index.ts
@@ -4,7 +4,6 @@ export * from './useOrchestratorConfig';
 export * from './useOrchestratorTheme';
 export * from './usePolicy';
 export * from './DataFetchHooks';
-export * from './useSubscriptionActions';
 export * from './useDataDisplayParams';
 export * from './useShowToastMessage';
 export * from './useStoredTableConfig';
