+++ packages/orchestrator-ui-components/src/types/types.ts
@@ -518,7 +518,7 @@ export type OrchestratorConfig = {
     graphqlEndpointCore: string;
     engineStatusEndpoint: string;
     processesEndpoint: string;
-    subscriptionActionsEndpoint: string;
+    subscriptionActionsEndpoint: string; //TODO: Remove
     subscriptionProcessesEndpoint: string;
     workflowInformationLinkUrl: string;
     authActive: boolean;
@@ -531,3 +531,23 @@ export enum ColorModes {
     LIGHT = 'LIGHT',
     DARK = 'DARK',
 }
+
+export interface SubscriptionAction {
+    name: string;
+    description: string;
+    reason?: string;
+    usable_when?: string[];
+    locked_relations?: string[];
+    unterminated_parents?: string[];
+    unterminated_in_use_by_subscriptions?: string[];
+    status?: string;
+    action?: string;
+}
+
+export type SubscriptionActions = {
+    reason?: string;
+    locked_relations?: string[];
+    modify: SubscriptionAction[];
+    terminate: SubscriptionAction[];
+    system: SubscriptionAction[];
+};
